% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-utils.R
\name{add_card_to_data}
\alias{add_card_to_data}
\title{Add a card to cards data}
\usage{
add_card_to_data(data, card)
}
\arguments{
\item{data}{Cards data list}

\item{card}{Card list to add}
}
\value{
Updated cards data
}
\description{
Add a card to cards data
}
\keyword{internal}
