% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_init}
\alias{config_init}
\title{Initialize config.yaml with default template}
\usage{
config_init(force = FALSE)
}
\arguments{
\item{force}{If TRUE, overwrite existing config file}
}
\value{
Invisible NULL
}
\description{
Initialize config.yaml with default template
}
