% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_set}
\alias{config_set}
\title{Set configuration values}
\usage{
config_set(type = NULL, key = NULL, value = NULL)
}
\arguments{
\item{type}{One of: "actor", "set", "room", "prop"}

\item{key}{The key to set}

\item{value}{The value to set}
}
\value{
Invisible NULL
}
\description{
Set configuration values
}
