% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-utils.R
\name{create_empty_cards}
\alias{create_empty_cards}
\title{Create an empty cards data structure}
\usage{
create_empty_cards()
}
\value{
List with empty cards structure
}
\description{
Create an empty cards data structure
}
\keyword{internal}
