% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_static_hanzi.R
\name{draw_static_hanzi}
\alias{draw_static_hanzi}
\title{Show static image of stroke order of Chinese character}
\usage{
draw_static_hanzi(hanzi, width, align, background_color)
}
\arguments{
\item{hanzi}{A required character string of one Chinese character (e.g., "你"); no default}

\item{width}{A character string for width of figure (e.g., "50\%" or "50px"); defaults to "100\%"}

\item{align}{A character string for alignment of figure (i.e., "center" or "left"); defaults to "center"}

\item{background_color}{A character string for figure color in CSS format (i.e., color name, HEX value, RGB value, or RGBA value)}
}
\value{
HTML+CSS code to generate stroke animation of specified Chinese character
}
\description{
This function generates the injectable HTML+CSS code to show a static image
of the stroke order of the specified Chinese character. If preparing a markdown
document in RStudio, set code chunk arguments as follows:
results='asis', message=F
}
