% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_anki_tsv}
\alias{export_anki_tsv}
\title{Export to Anki TSV}
\usage{
export_anki_tsv(cards, out_dir)
}
\arguments{
\item{cards}{Cards tibble}

\item{out_dir}{Output directory}
}
\description{
Export to Anki TSV
}
\keyword{internal}
