% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_csv}
\alias{export_csv}
\title{Export to CSV}
\usage{
export_csv(cards, out_dir)
}
\arguments{
\item{cards}{Cards tibble}

\item{out_dir}{Output directory}
}
\description{
Export to CSV
}
\keyword{internal}
