% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_markdown}
\alias{export_markdown}
\title{Export to Markdown}
\usage{
export_markdown(cards, out_dir)
}
\arguments{
\item{cards}{Cards tibble}

\item{out_dir}{Output directory}
}
\description{
Export to Markdown
}
\keyword{internal}
