% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{extract_component_chars}
\alias{extract_component_chars}
\title{Extract component characters from normalized components}
\usage{
extract_component_chars(components)
}
\arguments{
\item{components}{Normalized components list}
}
\value{
Character vector of component characters
}
\description{
Extract component characters from normalized components
}
\keyword{internal}
