% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_cards_file}
\alias{find_cards_file}
\title{Find the cards.yaml file}
\usage{
find_cards_file(must_exist = TRUE)
}
\arguments{
\item{must_exist}{If TRUE, abort if file doesn't exist}
}
\value{
Path to cards.yaml or NULL
}
\description{
Looks for cards.yaml in the current directory, then in inst/data/
}
\keyword{internal}
