% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{find_config_file}
\alias{find_config_file}
\title{Find the config.yaml file}
\usage{
find_config_file(must_exist = TRUE)
}
\arguments{
\item{must_exist}{If TRUE, abort if file doesn't exist}
}
\value{
Path to config.yaml or NULL
}
\description{
Looks for config.yaml in the same directory as cards.yaml
}
\keyword{internal}
