% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_date}
\alias{format_date}
\title{Format a date}
\usage{
format_date(time = Sys.time())
}
\arguments{
\item{time}{POSIXct time object}
}
\value{
Date string (YYYY-MM-DD)
}
\description{
Format a date
}
\keyword{internal}
