% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_timestamp}
\alias{format_timestamp}
\title{Format a timestamp}
\usage{
format_timestamp(time = Sys.time())
}
\arguments{
\item{time}{POSIXct time object}
}
\value{
ISO 8601 formatted string
}
\description{
Format a timestamp
}
\keyword{internal}
