% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-utils.R
\name{format_tone}
\alias{format_tone}
\title{Format tone for display}
\usage{
format_tone(tone, include_color = TRUE)
}
\arguments{
\item{tone}{Tone number (1-5)}

\item{include_color}{If TRUE, use cli colors in output}
}
\value{
Formatted tone string
}
\description{
Create accessible tone representation with number, shape, and pattern
}
\examples{
format_tone(3)  # "Tone 3 (dip \\/)"
}
