% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_actor}
\alias{get_actor}
\title{Get actor for a given initial}
\usage{
get_actor(initial, config = NULL)
}
\arguments{
\item{initial}{Pinyin initial (e.g., "sh", "b")}

\item{config}{Config data. If NULL, reads from file}
}
\value{
Actor name or NULL if not found
}
\description{
Get actor for a given initial
}
