% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-utils.R
\name{get_cards_tibble}
\alias{get_cards_tibble}
\title{Get all cards as a tibble}
\usage{
get_cards_tibble(path = NULL)
}
\arguments{
\item{path}{Path to cards.yaml file}
}
\value{
Tibble with card data
}
\description{
Get all cards as a tibble
}
