% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_config_init_path}
\alias{get_config_init_path}
\title{Get the default config.yaml path for initialization}
\usage{
get_config_init_path()
}
\value{
Path for new config.yaml
}
\description{
Returns the path where config.yaml should be created
}
\keyword{internal}
