% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_init_path}
\alias{get_init_path}
\title{Get the default cards.yaml path for initialization}
\usage{
get_init_path()
}
\value{
Path for new cards.yaml
}
\description{
Returns the path where cards.yaml should be created
}
\keyword{internal}
