% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_prop}
\alias{get_prop}
\title{Get prop meaning for a given component}
\usage{
get_prop(component, config = NULL)
}
\arguments{
\item{component}{Component character}

\item{config}{Config data. If NULL, reads from file}
}
\value{
Prop meaning or NULL if not found
}
\description{
Get prop meaning for a given component
}
