% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_room}
\alias{get_room}
\title{Get room for a given tone}
\usage{
get_room(tone, config = NULL)
}
\arguments{
\item{tone}{Tone number (1-5)}

\item{config}{Config data. If NULL, reads from file}
}
\value{
Room name or NULL if not found
}
\description{
Get room for a given tone
}
