% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{get_set}
\alias{get_set}
\title{Get set (location) for a given final}
\usage{
get_set(final, config = NULL)
}
\arguments{
\item{final}{Pinyin final (e.g., "ao", or null final for standalone initials)}

\item{config}{Config data. If NULL, reads from file}
}
\value{
Set name or NULL if not found
}
\description{
Get set (location) for a given final
}
