% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-utils.R
\name{get_tone_color_name}
\alias{get_tone_color_name}
\title{Get tone color name}
\usage{
get_tone_color_name(tone)
}
\arguments{
\item{tone}{Tone number (1-5)}
}
\value{
Color name
}
\description{
Convert tone number to color name (for accessibility)
}
\keyword{internal}
