% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-utils.R
\name{get_tone_pattern}
\alias{get_tone_pattern}
\title{Get tone pattern}
\usage{
get_tone_pattern(tone)
}
\arguments{
\item{tone}{Tone number (1-5)}
}
\value{
Pattern string
}
\description{
Convert tone number to visual pattern
}
\examples{
get_tone_pattern(1)  # "---"
get_tone_pattern(4)  # "\\"
}
