% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-utils.R
\name{get_tone_shape}
\alias{get_tone_shape}
\title{Get tone shape name}
\usage{
get_tone_shape(tone)
}
\arguments{
\item{tone}{Tone number (1-5)}
}
\value{
Shape name (singing, unsure, zombie, assertive, contrarian)
}
\description{
Convert tone number to accessible shape name using Mandarin Blueprint terminology
}
\examples{
get_tone_shape(1)  # "singing"
get_tone_shape(3)  # "zombie"
}
