% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{hanzi_actors}
\alias{hanzi_actors}
\title{List all actors in mnemonic system}
\usage{
hanzi_actors(show_usage = FALSE)
}
\arguments{
\item{show_usage}{If TRUE, also show which cards use each actor}
}
\value{
Invisible NULL
}
\description{
Display all actors with their associated Pinyin initials
}
