% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{hanzi_config}
\alias{hanzi_config}
\title{Manage mnemonic system configuration}
\usage{
hanzi_config(subcommand = "show", ...)
}
\arguments{
\item{subcommand}{One of: "init", "show", "validate", "set"}

\item{...}{Additional arguments passed to subcommands}
}
\value{
Invisible NULL
}
\description{
Main entry point for config subcommands
}
