% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{hanzi_filter}
\alias{hanzi_filter}
\title{Filter cards by criteria}
\usage{
hanzi_filter(args = character())
}
\arguments{
\item{args}{Command line arguments with filter options}
}
\value{
Invisible NULL
}
\description{
Filter cards by criteria
}
