% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{hanzi_init}
\alias{hanzi_init}
\title{Initialize a new cards.yaml file}
\usage{
hanzi_init(path = NULL, force = FALSE)
}
\arguments{
\item{path}{Path where cards.yaml should be created. If NULL, uses default location.}

\item{force}{If TRUE, overwrite existing file without prompting}
}
\value{
Invisible path to created file
}
\description{
Creates a new cards.yaml file with example cards
}
