% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnemonic.R
\name{hanzi_mnemonic}
\alias{hanzi_mnemonic}
\title{Show mnemonic information for a card}
\usage{
hanzi_mnemonic(char)
}
\arguments{
\item{char}{Character to show mnemonic for}
}
\value{
Invisible NULL
}
\description{
Quick-view of just the mnemonic information (actor, set, room, scene)
}
