% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{hanzi_props}
\alias{hanzi_props}
\title{List all props in mnemonic system}
\usage{
hanzi_props(show_usage = FALSE, limit = NULL)
}
\arguments{
\item{show_usage}{If TRUE, also show which cards use each prop}

\item{limit}{Maximum number of props to display (NULL for all)}
}
\value{
Invisible NULL
}
\description{
Display all props (component meanings) in alphabetical order
}
