% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{hanzi_search}
\alias{hanzi_search}
\title{Search cards by query}
\usage{
hanzi_search(queries)
}
\arguments{
\item{queries}{Character vector of search terms}
}
\value{
Invisible NULL
}
\description{
Full-text search across all fields
}
