% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{hanzi_sets}
\alias{hanzi_sets}
\title{List all sets in mnemonic system}
\usage{
hanzi_sets(show_usage = FALSE)
}
\arguments{
\item{show_usage}{If TRUE, also show which cards use each set}
}
\value{
Invisible NULL
}
\description{
Display all sets (locations) with their associated Pinyin finals
}
