% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show.R
\name{hanzi_show}
\alias{hanzi_show}
\title{Show detailed information about a card}
\usage{
hanzi_show(char)
}
\arguments{
\item{char}{Character to show}
}
\value{
Invisible NULL
}
\description{
Show detailed information about a card
}
