% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is_hanzi}
\alias{is_hanzi}
\title{Check if a character is valid Hanzi}
\usage{
is_hanzi(char)
}
\arguments{
\item{char}{Character to check}
}
\value{
Logical
}
\description{
Basic check if character is in CJK unicode range
}
\keyword{internal}
