% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{normalize_components}
\alias{normalize_components}
\title{Normalize component to list format}
\usage{
normalize_components(components)
}
\arguments{
\item{components}{Vector or list of components}
}
\value{
List of components in normalized format
}
\description{
Converts components to a standardized format that supports both
v1.0 (string) and v2.0 (object with char and meaning) formats
}
\keyword{internal}
