% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parse_all_options}
\alias{parse_all_options}
\title{Parse all options with a given prefix}
\usage{
parse_all_options(args, option)
}
\arguments{
\item{args}{Character vector of arguments}

\item{option}{Option name (e.g., "--tag")}
}
\value{
Character vector of all values for this option
}
\description{
Parse all options with a given prefix
}
\keyword{internal}
