% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli.R
\name{parse_option}
\alias{parse_option}
\title{Parse option from command line arguments}
\usage{
parse_option(args, option, default = NULL)
}
\arguments{
\item{args}{Character vector of arguments}

\item{option}{Option name (e.g., "--out")}

\item{default}{Default value if option not found}
}
\value{
Option value or default
}
\description{
Parse option from command line arguments
}
\keyword{internal}
