% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-utils.R
\name{parse_tone_from_pinyin}
\alias{parse_tone_from_pinyin}
\title{Parse tone from pinyin}
\usage{
parse_tone_from_pinyin(pinyin)
}
\arguments{
\item{pinyin}{Pinyin string with or without tone marks}
}
\value{
Tone number (1-5) or NA
}
\description{
Extract tone number from pinyin with tone marks
}
\keyword{internal}
