% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-utils.R
\name{read_cards}
\alias{read_cards}
\title{Read cards from YAML file}
\usage{
read_cards(path = NULL)
}
\arguments{
\item{path}{Path to cards.yaml file. If NULL, uses find_cards_file()}
}
\value{
List with cards data
}
\description{
Read cards from YAML file
}
