% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{read_config}
\alias{read_config}
\title{Read config from YAML file}
\usage{
read_config(path = NULL)
}
\arguments{
\item{path}{Path to config.yaml file. If NULL, uses find_config_file()}
}
\value{
List with config data
}
\description{
Read config from YAML file
}
