% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_line}
\alias{read_line}
\title{Read a line from stdin (works in both interactive and non-interactive mode)}
\usage{
read_line(prompt = "")
}
\arguments{
\item{prompt}{Prompt to display}
}
\value{
String input from user
}
\description{
Read a line from stdin (works in both interactive and non-interactive mode)
}
\keyword{internal}
