% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{truncate_text}
\alias{truncate_text}
\title{Truncate text to maximum length}
\usage{
truncate_text(text, max_len = 50, suffix = "...")
}
\arguments{
\item{text}{Character vector}

\item{max_len}{Maximum length}

\item{suffix}{Suffix to add if truncated}
}
\value{
Truncated text
}
\description{
Truncate text to maximum length
}
\keyword{internal}
