% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tone-utils.R
\name{validate_tone}
\alias{validate_tone}
\title{Validate tone value}
\usage{
validate_tone(tone)
}
\arguments{
\item{tone}{Tone value to validate}
}
\value{
TRUE if valid, FALSE otherwise
}
\description{
Validate tone value
}
\keyword{internal}
