% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-utils.R
\name{write_cards}
\alias{write_cards}
\title{Write cards to YAML file}
\usage{
write_cards(data, path = NULL)
}
\arguments{
\item{data}{List with cards data}

\item{path}{Path to cards.yaml file}
}
\value{
Invisible NULL
}
\description{
Write cards to YAML file
}
