% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config-utils.R
\name{write_config}
\alias{write_config}
\title{Write config to YAML file}
\usage{
write_config(data, path = NULL)
}
\arguments{
\item{data}{List with config data}

\item{path}{Path to config.yaml file}
}
\value{
Invisible NULL
}
\description{
Write config to YAML file
}
