% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hanzi-package.R
\docType{package}
\name{hanziR-package}
\alias{hanziR}
\alias{hanziR-package}
\title{hanziR: Terminal-Based Hanzi Learning Card Manager}
\description{
A minimal R package providing a console app for managing Chinese character (Hanzi) learning cards with accessible tone cues and multiple export formats. Features include YAML-based storage, full-text search, filtering, and export to Markdown, CSV, and Anki formats.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lddurbin/hanziR}
  \item Report bugs at \url{https://github.com/lddurbin/hanziR/issues}
}

}
\author{
\strong{Maintainer}: Lee Durbin \email{hello@leedurbin.co.nz}

}
\keyword{internal}
